fix: users fields
