+++ examples/app-crm/src/components/kanban-checklist-header.tsx
@@ -1,20 +1,14 @@
-import { Space, Tag, Typography } from "antd";
+import { Typography } from "antd";
 
 import { Task } from "../interfaces/graphql";
 
 type Props = {
-    users?: Task["users"];
+    checklist?: Task["checklist"];
 };
 
-export const KanbanUsersHeader = ({ users = [] }: Props) => {
-    if (users.length > 0) {
-        return (
-            <Space>
-                {users?.map((user) => (
-                    <Tag key={user.id}>{user.name}</Tag>
-                ))}
-            </Space>
-        );
+export const KanbanChecklistHeader = ({ checklist = [] }: Props) => {
+    if (checklist.length > 0) {
+        return <div>Not implemented</div>;
     }
 
     return <Typography.Link>Assign to users</Typography.Link>;
