--- examples/app-crm/src/components/kanban-user-header.tsx
@@ -1,15 +1,0 @@
-import { Typography } from "antd";
-
-import { Task } from "../interfaces/graphql";
-
-type Props = {
-    checklist?: Task["checklist"];
-};
-
-export const KanbanChecklistHeader = ({ checklist = [] }: Props) => {
-    if (checklist.length > 0) {
-        return <div>Not implemented</div>;
-    }
-
-    return <Typography.Link>Add checklist</Typography.Link>;
-};
